Move user management types
