+++ workos/types/user_management/password_hash_type.py
@@ -1,0 +1,4 @@
+from typing import Literal
+
+
+PasswordHashType = Literal["bcrypt", "firebase-scrypt", "ssha"]
