+++ workos/types/user_management/user_management_provider_type.py
@@ -1,0 +1,6 @@
+from typing import Literal
+
+
+UserManagementProviderType = Literal[
+    "authkit", "AppleOAuth", "GitHubOAuth", "GoogleOAuth", "MicrosoftOAuth"
+]
